Add a few new shows
LotR: The Rings of Power was solid. Never seen any of the films, but this was interesting.

The Peripheral was intriguing. Had some futuristic tech sci-fi thriller kind of vibes. Definitely recommend. Amazon is only doing weekly releases for this one though.

The Watcher.. what a mind bending show based on a true story. If you like true crime and mysteries -- this is the one for you.